Fix logo resizing and positionting.
On the iPhone, things got all mucked up. This is becuase the size of the
blog title changed, but not the image. So add media selectors to resize the
image, too.

While at it, reverse the order of the logo and the title and eliminate the
absolute positioning of the title. This makes it work properly on Chrome.