Undline linklog URLs again.
But use a bottom border to get nice spacing instead of the
sucky default underline spacing, which tends to run through
descenders. Works on hover, too, but the text color changes
slowly, while the border color change is fast. Not sure what
makes the text color change take 500ms or so. Will figure it
out later.